requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
aiohttp>=3.8.0
pyyaml>=6.0
scikit-learn>=1.3.0
asyncio-mqtt>=0.13.0
websockets>=11.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
